requirements.txt
lxml==4.9.1
pycryptodome==3.15.0
requests==2.27.1
Pillow
pyzbar
qrcode
